Fix collections checksum hook
The checksum hook for collections was not properly substituting values
as it was accessing the root values for the yaml file.

This patch makes it so it uses the YQ_ARGS available in other places in
order to access the proper subpackage when dealing with collections.

This pathc also changes the sed separator char to an @ as we currently
use / and that collides if there is an url as a value. It should be more
difficult to find an @ in the values..I hope

Signed-off-by: Itxaka <igarcia@suse.com>