Add github_tag strategy
Currently if using the "non-named" github strategy, autobum would use
the github{owner,repo} to check the new version but the original package
version would not be stored anywhere, thus tyieng the package version
with the underlying artifacts.

This intriduces the github_tag strategy which will store the original
remote source version into labels.github.tag for further use by packages

It also adds a test package for it and sets pipefail to exit so we can
properly fail on testing when dealing with pipes.

All current uses of the autobumper should not be affected by this change

Signed-off-by: Itxaka <igarcia@suse.com>